fix(Removable): optimize component rendering with indent
